+++ include/apps/watchfaces/OswAppWatchfaceFitnessAnalog.h
@@ -38,8 +38,8 @@ class OswAppWatchfaceFitnessAnalog : public OswAppV2 {
     unsigned screen = 0;
 
     void showFitnessTracking(OswHal* hal);
-    void drawWatchFace(OswHal* hal, uint32_t hour, uint32_t minute, uint32_t second, bool afterNoon);
-    void drawDateFace(OswHal* hal, uint32_t hour, uint32_t minute, uint32_t second, bool afterNoon);
+    void drawWatchFace(OswHal* hal, uint8_t hour, uint8_t minute, uint8_t second, bool afterNoon);
+    void drawDateFace(OswHal* hal, uint8_t hour, uint8_t minute, uint8_t second, bool afterNoon);
 
 #ifdef GIF_BG
     OswAppGifPlayer* bgGif = nullptr;
